+++ app/src/main/java/org/simple/clinic/home/overdue/OverdueModel.kt
@@ -3,18 +3,21 @@ package org.simple.clinic.home.overdue
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 import org.simple.clinic.facility.Facility
+import org.simple.clinic.home.overdue.PendingListState.SEE_LESS
 
 @Parcelize
 data class OverdueModel(
     val facility: Facility?,
-    val overdueAppointmentSections: OverdueAppointmentSections?
+    val overdueAppointmentSections: OverdueAppointmentSections?,
+    val pendingListState: PendingListState
 ) : Parcelable {
 
   companion object {
     fun create(): OverdueModel {
       return OverdueModel(
           facility = null,
-          overdueAppointmentSections = null
+          overdueAppointmentSections = null,
+          pendingListState = SEE_LESS
       )
     }
   }
@@ -39,5 +42,9 @@ data class OverdueModel(
       overdueAppointmentSections: OverdueAppointmentSections
   ): OverdueModel {
     return copy(overdueAppointmentSections = overdueAppointmentSections)
+  }
+
+  fun pendingListStateChanged(state: PendingListState): OverdueModel {
+    return copy(pendingListState = state)
   }
 }